A parameter declared with a NEGATIVE low bound reads its declared range (§2 🆕 L ⓩ residue, §4.5.448)
Bug (silent-wrong, both oracles): `localparam logic [3:-4] A = 8'h3C; localparam L =
{A[3:0], A[3:0]};` printed `cc` where both oracles print `33`, and beside it the bare
select `A[3:0]` and the bit-select `A[-1]` were honest-loud ("the part-select `A[…]`
has no constant-fold arm"). One root: `DeclRange`'s `lo` was a `u32`, so
`param_decl_range_opt` DECLINED a negative bound outright and every consumer fell back
to reading the normalised `[w-1:0]` storage positionally.

Mechanism: `lo` widened to `i64` across the carriers that spell the alias —
`DeclRange` itself (and with it `pkg_const_range` and `enum_label_prepass`), the two
spelled-out maps `param_range` / `hier_param_range`, `param_decl_range_opt` /
`bind_param_range` / `param_sel_range`, and the const lane's `const_select_base` /
`select_base_at_declared` / `select_idx_in_declared_range`. `norm_offset_for_range`
and `const_norm_bit` were ALREADY i64 — the offset arithmetic for a negative lo
existed and was exercised by the packed-element path, which is exactly why the runtime
lane was right and the fold lane was not; their call sites only shed an `i64::from`.

The decline this replaced was right for what it guarded: its comment records a
measured regression from writing a CLAMPED lo (an ascending `[-2:3]` recorded
`(0, 6, true)` and turned two correct cells into silent-wrongs). A truthful i64 is not
that, and the keys that GAIN an entry are exactly the ones with a negative bound, so
every non-negative declaration is untouched — the `[11:4]`, `[7:0]` and `[0:7]`
controls are byte-identical in all three tools.

Census: 33 cells, 12 WRONG→CORRECT, 1 LOUD→CORRECT, 0 regressions. Every declarer
spelling moves together (localparam, `parameter`, a package constant read bare and
scoped, an OVERRIDDEN header parameter, a generate-scope one, and a hierarchical
`u.A[3:0]` through the persistent twin map), and so does every consumer the review
attacked: a concatenation, a bare select, a bit-select at a negative index, both
bounds negative, the ascending spelling (iverilog-only — verilator refuses `logic
[-4:3]`), a const-function body, a RANGE BOUND and a REPLICATION COUNT (`wire
[A[3:0]:0] w` went from a silent 1-bit net to both oracles' 4, `{A[3:0]{1'b1}}` from
`fff` to `7`).

Recorded, three separate roots left on the axis: a negative select BOUND (`A[0:-2]`)
false-louds because `const_bound_u32` folds it unsigned; a >64-bit negative-LSB base
stays positional (a different lane reads `param_range` directly and
`select_base_at_declared` refuses `dwidth > 64`); an explicit `[m:l]` PART select of a
NET with a negative low bound is loud by its own gate. Corpus demand is ZERO.

Review (batch, §4.5.446–448): 2 lenses run directly, 0 BLOCKING. Gate 7,202 green,
clippy and fmt clean, corpus 10/10 with no drift, examples byte-identical PRE vs POST,
staged == one-shot, product build (`--no-default-features`) green.

Batch docs: ROADMAP §2 (🆕 O and 🆕 L ⓩ → residues, new rows 🆕 P / 🆕 Q), §3 ⑤ⓕ
residue with its oracle count corrected, §5.2 queue; ARCHIVE §4.5.446–448; CHANGELOG;
REMAINING_WORK; manual 003; ENGINEERING_RULES (three lessons).

+++ crates/elaborate/src/const_array.rs
@@ -233,7 +233,7 @@ impl Elaborator<'_> {
                 ast::ExprKind::Ident(_) | ast::ExprKind::PkgScoped { .. }
             ) {
                 let (lo, w, asc) = self.param_sel_range(arg)?;
-                let (lo, hi) = (i64::from(lo), i64::from(lo) + i64::from(w) - 1);
+                let (lo, hi) = (lo, lo + i64::from(w) - 1);
                 (vec![if asc { (lo, hi) } else { (hi, lo) }], 1, 0)
             } else {
                 return None;
